// workspaces/electron-app/main/models/channel-name.const.ts
export enum ChannelName {
  DOWNLOAD_LOGS = 'downloadLogs',
  SHOW_OPEN_DIALOG = 'showOpenDialog',
  SHOW_SAVE_DIALOG = 'showSaveDialog',
  GET_APP_PATH = 'getAppPath',
  GET_PATH = 'getPath',
  DIRNAME = 'dirname',
  PATH_JOIN = 'pathJoin',
  GET_NATIVE_THEME = 'getNativeTheme',
  CREATE_AUTH_WINDOW = 'createAuthWindow',
  CREATE_AUTH_LOGOUT_WINDOW = 'createAuthLogoutWindow',
}